Updated badges to be more mobile responsive
