+++ src/components/Badges/Badges.module.scss
@@ -81,13 +81,3 @@
     }
   }
 }
-
-.tabs {
-  :global(.ui.menu) {
-    overflow: auto;
-  }
-
-  :global(.ui.bottom.segment) {
-    padding: 0.25em;
-  }
-}
