+++ src/components/Badges/elements/BadgeDetail/BadgeDetail.module.scss
@@ -1,6 +1,7 @@
 .title {
   font-weight: bold;
   font-size: 24px;
+  color: var(--contrast);
 }
 
 .rematch {
@@ -33,6 +34,7 @@
         flex-flow: column nowrap;
         align-items: center;
         gap: 5px;
+        color: var(--contrast);
 
         > img {
           height: 60px;
@@ -61,6 +63,7 @@
         flex-flow: column nowrap;
         align-items: center;
         gap: 3px;
+        color: var(--contrast);
 
         > span {
           font-size: 12px;
