+++ src/components/Badges/elements/index.ts
@@ -1,0 +1,3 @@
+export { default as BadgeDetail } from './BadgeDetail/BadgeDetail';
+export { default as BadgeEditor } from './BadgeEditor/BadgeEditor';
+export { default as BadgePage } from './BadgePage/BadgePage';
